refactor: 자동차 이름 검증 Car 클래스에서 하도록 수정
