+++ src/GameSimulator/constants.js
@@ -1,11 +1,1 @@
-export const CAR_NAME_LENGTH = {
-  MIN: 1,
-  MAX: 5,
-};
-
-export const CAR_NAME_ERROR_MESSAGE = {
-  LESS_THAN_MIN: `자동차 이름은 최소 ${CAR_NAME_LENGTH.MIN}글자 이상 입니다.`,
-  OVER_THAN_MAX: `자동차 이름은 최대 ${CAR_NAME_LENGTH.MAX}글자 입니다.`,
-};
-
 export const MAX_ROUNDS = 5;
